backend/foodgram/requirements.txt
asgiref==3.5.1
atomicwrites==1.4.0
attrs==21.4.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4
cryptography==37.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.13
django-filter==21.1
django-import-export==2.8.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
et-xmlfile==1.1.0
idna==3.3
importlib-metadata==1.7.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.1.1
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tablib==3.2.1
tomli==2.0.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
